V1.1
Fixed all the docking ports.
Re-pass over the Science lab - but it's legs also have issues like the
Ranger. For now, as long as you don't launch it with it'scaled
legs deployed it seems fine.
Added KIS container space to the Ranger and Lander (if you have KIS
installed).
Re-balanced the Ranger and Lander centre of mass and centre of lift and
engine thrusters to hopefully make them much more stable.